rename and move trajectory analysis scripts
